[Gutenberg] New file for GutenbergRollout struct
